+++ Changes.md
@@ -3,6 +3,8 @@ Version Information
 
 Version 403.1.2 - TBR
 --------------------------
+1. Add the ability not to show 'My courses' in the breadcrumb when it is not used as the start page for the user.  See the
+  'breadcrumbdisplaymycourses' setting in on the 'Header' settings tab under 'Breadcrumb'.
 
 Version 403.1.1 - 12/10/23
 --------------------------